Turn off merge applicability check by default
Window which going to be merged; especially forwarded from other hosts,
is expected not to fit any active/waiting windows frequently.

Assume the case figured below. Startup time of each process in host A
and B differs, so generated window's start time gets differ. In this
case host B (or A) windows can't get merged in host A (or B) windows.

    time   | 0123456789
    host A |  |--|--|
    host B |    |--|--|

So this fix is turning merge applicability check off for windows
generated by WindowScheduler in default.

Related-To: 969446d2474f64a0c2a5536d5708eebf27e5602c